requirements.txt
beets
google-genai
